Refactor KL divergence and prefix cache branching to kits, update all Qwen3-Next tests
# Conflicts:
#	test/registered/4-gpu-models/test_qwen3_next_models_mtp.py
